refactor: move WorkspaceActions context
containers/WorkspaceActions -> contexts/WorkspaceActions
