+++ packages/dashboard-frontend/src/contexts/WorkspaceActions/__tests__/index.spec.tsx
@@ -15,9 +15,9 @@ import { Action, Store } from 'redux';
 import { Provider } from 'react-redux';
 import { render, RenderResult, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import WorkspaceActionsProvider from '..';
+import WorkspaceActionsProvider from '../Provider';
 import { WorkspaceAction } from '../../../services/helpers/types';
-import { ActionContextType, WorkspaceActionsConsumer } from '../context';
+import { ActionContextType, WorkspaceActionsConsumer } from '..';
 import { FakeStoreBuilder } from '../../../store/__mocks__/storeBuilder';
 import { createFakeCheWorkspace } from '../../../store/__mocks__/workspace';
 import { ActionCreators } from '../../../store/Workspaces';
